pyproject.toml
[project]
name = "knowledge-bot"
version = "0.1.0"
description = "A knowledge bot using Groq(openai/gpt-oss-120b) and Qdrant"
authors = [{name = "Jonindo Pasaribu"}]
readme = "README.md"
requires-python = ">=3.9"

[tool.poetry]
packages = [{include = "pipeline"}, {include = "chatbot"}, {include = "common"}]

[tool.poetry.dependencies]
python = "^3.9"
qdrant-client = "^1.15.1"
google-generativeai = "^0.8.5"
python-dotenv = "^1.1.1"
pypdf = "^6.0.0"
tqdm = "^4.67.1"
langchain-huggingface = "^0.3.1"
langchain_community = "^0.3.27"
langchain-qdrant = "^0.2.0"
langchain-groq = "^0.3.7"
langchain = "^0.3.27"
sentence-transformers = "^5.1.0"
termcolor = "^3.1.0"

[tool.poetry.scripts]
ingest = "pipeline.ingest_documents:ingest_documents"
chatbot = "chatbot.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
